Moved controllers to a class library so they can be used from other backends.
