+++ src/ConfigUtil/project.json
@@ -1,8 +1,8 @@
 {
   "dependencies": {
-    "Microsoft.NETCore.App": "1.0.0",
-    "Microsoft.AspNetCore.Mvc": "1.0.0",
-    "Microsoft.AspNetCore.Server.Kestrel": "1.0.0",
+    "Microsoft.NETCore.App": "1.0.1",
+    "Microsoft.AspNetCore.Mvc": "1.0.1",
+    "Microsoft.AspNetCore.Server.Kestrel": "1.0.1",
     "Microsoft.Extensions.Configuration.EnvironmentVariables": "1.0.0",
     "Microsoft.Extensions.Configuration.FileExtensions": "1.0.0",
     "Microsoft.Extensions.Configuration.Json": "1.0.0",
@@ -11,7 +11,8 @@
     "Microsoft.Extensions.Logging.Debug": "1.0.0",
     "Microsoft.Extensions.Options.ConfigurationExtensions": "1.0.0",
     "Microsoft.AspNetCore.StaticFiles": "1.0.0",
-    "OSVR.Config": "1.0.0-*"
+    "OSVR.Config": "1.0.0-*",
+    "OSVR.Config.Controllers": "1.0.0-*"
   },
 
   "frameworks": {
