+++ src/OSVR.Config.Controllers/project.json
@@ -1,8 +1,10 @@
-﻿{
+{
   "version": "1.0.0-*",
 
   "dependencies": {
-    "NETStandard.Library": "1.6.0"
+    "Microsoft.AspNetCore.Mvc": "1.0.1",
+    "NETStandard.Library": "1.6.0",
+    "OSVR.Config": "1.0.0-*"
   },
 
   "frameworks": {
